peak detecting alorithm reworked aka "adjusting threshold bug"
The "ignore approx x bp" is not affecting the detection algorithm in any way. The minimum threshold will be only used if the calculated "maximum peaks of the lower ones" is below 300 RFU.
Previous Version of peakAdaptM first scanned the data for peaks at a minimum threshold of 150 RFU, after this first detection some variables (maximum dist between peaks, peak prominence and threshold) will be adjusted and then the highest peak was removed if it was "far enough" away from the median of the other peaks.

Main problem of the new data was the, in comparsion to the other peaks, massive false signal. This caused the algorithm for adapting threshold to use a much to high adjusted threshold as intended.

This approach "from the borders to the middle" is replaced by a new version 
after the first scan the percentual difference between the median of all found peak heights, and the mean of all found peaks except the biggest. If the difference is high